+++ Source/Bogus/DataSets/Images.PlaceImg.cs
@@ -1,0 +1,18 @@
+﻿namespace Bogus.DataSets
+{
+   public enum PlaceImgFilter
+   {
+      Grayscale,
+      Sepia
+   }
+
+   public static class PlaceImgCategory
+   {
+      public const string Animals = "animals";
+      public const string Architecture = "arch";
+      public const string Nature = "nature";
+      public const string People = "people";
+      public const string Tech = "tech";
+      public const string Any = "any";
+   }
+}